spider/requirements.txt
mysql-connector-python
sqlalchemy
requests
lxml